Add recursive fuel required for part 2
Part 1 of the day 1 puzzle asked you to compute the fuel required for
the spacecraft to launch into space using the provided calculation.

For part 2, it was clearly asking for a recursive version of the
calculation where the result of the first time it's performed gets fed
into the next time, and so on until the input is <= 0.