Solve day2 part 1
This puzzle implements a simple "intcode computer" which reads through
the input in steps of 4. The value at each step is the "opcode" followed
by two pointers to the inputs and finally a pointer to where the result
of the operation should be stored.Solve day2 part 1

This puzzle implements a simple "intcode computer" which reads through
the input in steps of 4. The value at each step is the "opcode" followed
by two pointers to the inputs and finally a pointer to where the result
of the operation should be stored.